# Optimize migration which creates task for each job