--- a/db/migrate/20170120164058_create_task_for_each_job_and_transfer_attributes.rb
+++ b/db/migrate/20170120164058_create_task_for_each_job_and_transfer_attributes.rb
@@ -12,6 +12,9 @@ class CreateTaskForEachJobAndTransferAttributes < ActiveRecord::Migration[5.0]
     say_with_time("Deleting finished jobs older than 7 days") do
       Job.where("updated_on < ?", Time.now.utc - 7.days).where("state = 'finished'").delete_all
     end
+    say_with_time("Deleting finished tasks older than 7 days") do
+      MiqTask.where("updated_on < ?", Time.now.utc - 7.days).where("state = 'Finished'").delete_all
+    end
     say_with_time("Creating tasks associated with jobs") do
       Job.find_each do |job|
         job.create_miq_task(:status        => job.status.try(:capitalize),