There are several long running migrations observed on client's DB (`Job.count` around 1.3 millions)  : 
`CreateTaskForEachJobAndTransferAttributes`        - 4863 seconds
`CopyAgentIdToMiqServerIdInJobsTable`                  -  168 seconds
`CopyServerIdFromJobsToMiqTask`                           - 1645

In UI (Task management screen) we show only tasks for the last 6 days, so there is no need to do anything with jobs/task older than 7 days.

This PR:  changed exiting `CreateTaskForEachJobAndTransferAttributes` by:
- deleting finished `Job` older than 7 days first and creating task for each job after that.
- deleting finished `MiqTask` older than 7 days. This change will no speed-up migration, it is just clen-up and done t be consistent with `Job` deletion. 

on the same client's DB deleting old records from `Job` table took 16 seconds:
```
>> Job.where("updated_on < ?", t).where("state = 'finished'").delete_all
  SQL (16330.1ms)  DELETE FROM "jobs" WHERE (updated_on < '2018-04-14 16:05:23.679929') AND (state = 'finished')
=> 1232220
```

Depends on https://github.com/ManageIQ/manageiq/pull/17727 for backports