@Fryguy 2018-05-21
@kbrock please review for performance.