@jrafanie 2018-05-21
Looks good to me.  Do we care that some people will have their tasks/jobs tables cleared and others will not?  (they ran this migration previously)  

@kbrock can you review?