@yrudman 2018-05-30
I think originally we did not clean-up tasks since report were not shown if associated task removed.
It was fixed in https://github.com/ManageIQ/manageiq-ui-classic/pull/1965https://github.com/ManageIQ/manageiq-ui-classic/pull/1965 and https://github.com/ManageIQ/manageiq-ui-classic/pull/1488

7 days limit was picked-up because we are not showing in UI tasks with latest update older than 6 days:
![screen shot 2018-05-30 at 2 53 24 pm](https://user-images.githubusercontent.com/6556758/40741614-f89e658c-6419-11e8-9616-5e84b8d268ce.png)