@Fryguy 2018-05-31
Something feels off about deleting old tasks.  Or put another way, this feels like a "purge" of tasks, but the value being chosen is "7 days" and it only happens once during this migration and then never again.  I think we need @bascar and @chessbyte's opinion here.