@gtanzillo 2018-06-07
@jdeubel Regarding https://github.com/ManageIQ/manageiq-schema/pull/201#issuecomment-393623651, do you have a feel for whether purging jobs older than 7 days would have and impact on customers?