@jdeubel 2018-06-08
@gtanzillo If that is the max amount of time that a user can see then that is fine with me.