@jrafanie 2018-06-08
> @gtanzillo If that is the max amount of time that a user can see then that is fine with me.

yeah, I agree, if we choose what max time in the pull down for how far "back" you can go to see old tasks/jobs throughout the UI, there is very little risk to removing older records.