@Fryguy 2018-06-28
I would prefer we add purging for MiqTasks defaulting to 7 days in manageiq core, to avoid growing the table *after* the migration.  Then, this PR would be fine (or perhaps deleting old should be a separate PR dated one before this one, but if we need to backport, putting them in one PR is fine)