@kbrock 2018-08-02
I merged the other pruning PR associated with this one.

@Fryguy this good to go?