@yrudman 2018-08-07
@Fryguy as we discussed - it ot optimization PR which is editing existing migration and will be backported together with https://github.com/ManageIQ/manageiq/pull/17727