@simaishi 2018-09-11
Gaprindashvili backport details:

```
$ git log -1
commit 44ade48a30ce898b4d5a456fbcbd5f817d379133
Author: Jason Frey <fryguy9@gmail.com>
Date:   Tue Aug 7 12:26:31 2018 -0400

    Merge pull request #201 from yrudman/delete-old-job-while-creating-linked-task
    
    Optimize migration which creates task for each job
    (cherry picked from commit 8d48f1bedf9c42661ecb83ae3ab5ab9fb2b21545)
    
    https://bugzilla.redhat.com/show_bug.cgi?id=1618808
```